api/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
eval-type-backport==0.2.0
fastapi==0.115.2
filelock==3.16.1
frozenlist==1.4.1
h11==0.14.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
numpy==2.1.2
pillow==10.4.0
propcache==0.2.0
pyarrow==17.0.0
pydantic==2.9.2
pydantic-core==2.23.4
pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.2
shellingham==1.5.4
sniffio==1.3.1
starlette==0.39.2
tabulate==0.9.0
together==1.3.1
tqdm==4.66.5
typer==0.12.5
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
yarl==1.15.1
